+++ CHANGELOG.md
@@ -61,6 +61,7 @@
 - disallow passing invalid nodes as arguments to functions and operators
 - checking for unevaluated spread inside macros
 - checking for invalid symbols when defining a function through a macro
+- added a max macro unification depth
 
 ### Removed
 - removed unused `NodeType::Closure`